moved UI handling code into WeatherService from SearchBar
